+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## [Unreleased]
 ### Changed
 - Wrap all NextFlow output with real-time line-by-line log statements
+- Wrap custom comparison to log through NFTest logger
 - Add DEBUG lines about specific asserts
 
 ---